Canary_1b_v2/server/requirements-linux.txt
fastapi
uvicorn[standard]
numpy
soundfile
nemo_toolkit[asr]
